00_cyberteam搭建/Output/cyberteam-v4/backend/requirements.txt
# CyberTeam V4 Backend Dependencies

# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0  # PostgreSQL async driver

# Auth
python-jose>=3.3.0  # JWT
passlib>=1.7.4
bcrypt>=4.0.0

# Redis
redis>=5.0.0

# YAML
pyyaml>=6.0.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Logging
structlog>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.0.0
ruff>=0.1.0
